Device/Port: don't store baud rate in a "long" integer
All of our target platforms have a 32 bit "standard" integer.  We
surely don't need a 64 bit baud rate.

+++ src/Device/AndroidBluetoothPort.cpp
@@ -153,15 +153,15 @@ AndroidBluetoothPort::SetRxTimeout(unsigned Timeout)
   return true;
 }
 
-unsigned long
+unsigned
 AndroidBluetoothPort::GetBaudrate() const
 {
   // XXX implement K6-Bt commands?
   return 19200;
 }
 
-unsigned long
-AndroidBluetoothPort::SetBaudrate(unsigned long BaudRate)
+unsigned
+AndroidBluetoothPort::SetBaudrate(unsigned BaudRate)
 {
   // XXX implement K6-Bt commands?
   return BaudRate;
