+++ src/Device/TCPPort.hpp
@@ -73,8 +73,8 @@ public:
   bool Close();
 
   virtual bool SetRxTimeout(unsigned Timeout);
-  virtual unsigned long GetBaudrate() const;
-  virtual unsigned long SetBaudrate(unsigned long BaudRate);
+  virtual unsigned GetBaudrate() const;
+  virtual unsigned SetBaudrate(unsigned BaudRate);
   virtual bool StopRxThread();
   virtual bool StartRxThread();
   void ProcessChar(char c);
